+++ Makefile
@@ -44,6 +44,7 @@ clean: ## Has gradle clean it's build files
 	./gradlew clean
 	rm fatjar.jar || true
 	rm -r fatjar || true
+	rm -r gallifrey/*/bin || true
 
 .PHONY: fatjar
 fatjar: ## Make sure you build before you try to make the jar file
