+++ gallifrey/core/src/main/java/gallifrey/core/AntidoteInterface.java
