+++ gallifrey/core/src/main/java/gallifrey/core/Counter.java
