+++ gallifrey/core/src/main/java/gallifrey/core/GenericFunction.java
