+++ gallifrey/core/src/main/java/gallifrey/core/PositiveCounter.java
