+++ gallifrey/core/src/main/java/gallifrey/core/SharedMap.java
