+++ gallifrey/core/src/test/java/gallifrey/core/CounterTest.java
