+++ gallifrey/core/src/test/java/gallifrey/core/PositiveCounterTest.java
