+++ gallifrey/core/src/test/java/gallifrey/core/RegisterTest.java
