+++ gallifrey/frontend/src/test/java/gallifrey/frontend/SharedLinkedListSerializationTest.java
@@ -4,21 +4,19 @@ import org.junit.Test;
 import static org.junit.Assert.assertEquals;
 
 import java.io.*;
-import java.util.Random;
 
-import com.google.protobuf.ByteString;
+import gallifrey.core.SharedObject;
+import gallifrey.core.SharedLinkedList;
 
-import gallifrey.core.Counter;
-
-public class SharedObjectSerializationTest {
+public class SharedLinkedListSerializationTest {
     @Test
     public void test() {
         try {
-            Random rd = new Random();
-            byte[] random_bytes = new byte[10];
-            rd.nextBytes(random_bytes);
-            ByteString random_key = ByteString.copyFrom(random_bytes);
-            SharedObject obj = new SharedObject(new Counter(0), random_key);
+            SharedLinkedList<String> obj = new SharedLinkedList<String>("bye");
+
+            SharedLinkedList<String> temp = new SharedLinkedList<String>("hello");
+            SharedObject s = new SharedObject(temp);
+            obj.setNext(s);
             FileOutputStream fout = new FileOutputStream("file.txt");
             ObjectOutputStream oos = new ObjectOutputStream(fout);
             oos.writeObject(obj);
@@ -27,9 +25,11 @@ public class SharedObjectSerializationTest {
 
             FileInputStream fin = new FileInputStream("file.txt");
             ObjectInputStream ois = new ObjectInputStream(fin);
-            SharedObject obj2 = (SharedObject) ois.readObject();
-            assertEquals(obj.key.getKey(), obj2.key.getKey());
-
+            SharedLinkedList<String> obj2 = (SharedLinkedList<String>) ois.readObject();
+            assertEquals(obj.getData(), "bye");
+            assertEquals(obj.getData(), obj2.getData());
+            assertEquals(s.key, obj2.getNext().key);
+            assertEquals(obj.getNext().key, obj2.getNext().key);
             ois.close();
         } catch (IOException | ClassNotFoundException e) {
             e.printStackTrace();
