+++ gallifrey/frontend/src/test/java/gallifrey/frontend/SharedObjectSerializationTest.java
@@ -4,16 +4,22 @@ import org.junit.Test;
 import static org.junit.Assert.assertEquals;
 
 import java.io.*;
+import java.util.Random;
 
-public class SharedLinkedListSerializationTest {
+import com.google.protobuf.ByteString;
+
+import gallifrey.core.Counter;
+import gallifrey.core.SharedObject;
+
+public class SharedObjectSerializationTest {
     @Test
     public void test() {
         try {
-            SharedLinkedList<String> obj = new SharedLinkedList<String>("bye");
-
-            SharedLinkedList<String> temp = new SharedLinkedList<String>("hello");
-            SharedObject s = new SharedObject(temp);
-            obj.setNext(s);
+            Random rd = new Random();
+            byte[] random_bytes = new byte[10];
+            rd.nextBytes(random_bytes);
+            ByteString random_key = ByteString.copyFrom(random_bytes);
+            SharedObject obj = new SharedObject(new Counter(0), random_key);
             FileOutputStream fout = new FileOutputStream("file.txt");
             ObjectOutputStream oos = new ObjectOutputStream(fout);
             oos.writeObject(obj);
@@ -22,11 +28,9 @@ public class SharedLinkedListSerializationTest {
 
             FileInputStream fin = new FileInputStream("file.txt");
             ObjectInputStream ois = new ObjectInputStream(fin);
-            SharedLinkedList<String> obj2 = (SharedLinkedList<String>) ois.readObject();
-            assertEquals(obj.getData(), "bye");
-            assertEquals(obj.getData(), obj2.getData());
-            assertEquals(s.key, obj2.getNext().key);
-            assertEquals(obj.getNext().key, obj2.getNext().key);
+            SharedObject obj2 = (SharedObject) ois.readObject();
+            assertEquals(obj.key.getKey(), obj2.key.getKey());
+
             ois.close();
         } catch (IOException | ClassNotFoundException e) {
             e.printStackTrace();
